Updated elixir backend to allow cors. It's currently set to allow all origins.
Added activity service to client to query the backends.

added env.ts as the default enviornment.
Updated app.config.ts so it correctly overrwrites the env files depending on which back the user wants to use.

updated package.json so if someone just types npm run start, it lets them know that a backend needs to be selected.